Rename oop_library.py to library.py and update its usages
