+++ library.py
@@ -22,80 +22,28 @@ Classes:
     Message:
         The class for each message in a chat. Every instance is a separate message.
 
+    Chat:
+        The class for each chat to be formatted. Every instance is a separate chat.
+
 Functions:
-    clean_html(string: str) -> str:
-        Remove < and > to avoid rogue HTML tags.
+    process_chat(input_file: str, group_chat: bool, sender_name: str, chat_title: str, html_file_name: str, output_dir: str) -> None:
+        Process one chat completely.
 
-    simple_format_to_html(string: str) -> str:
-        Replace WhatsApp format characters with their HTML tags. Doesn't handle code blocks.
+    process_list_of_chats(list_of_chats: list) -> list:
+        Fully format a list of lists, where each sub-list is a set of arguments to be passed to process_chat().
 
-    complete_format_to_html(string: str) -> str:
-        Replace WhatsApp format characters with their HTML tags. Handles code blocks.
-
-    format_links(string: str) -> str:
-        Find all the links in the string and put them in <a> tags.
-
-    fully_format_message_content(string: str) -> str:
-        Take message content and fully format it properly.
-
-    add_attachments(attachment: str) -> str:
-        Create HTML tags to embed attachments.
-
-    write_text(chat_title: str, group_chat: bool):
-        Format and write all text in _chat.txt to HTML file.
-
-    move_attachment_files():
-        Move all attachment files that don't need to be converted.
-
-    extract_zip(input_file: str) -> bool:
-        Extract the zip file given by file_dir into the temp directory. Return True if successfully extracted.
-
-    fully_format_chat(group_chat: bool, sender_name: str, chat_title: str, html_file_name: str, output_dir: str):
-        Fully format an already exported chat.
-
-    process_single_chat(input_file: str, group_chat: bool, sender_name: str, chat_title: str, html_file_name: str, output_dir: str):
-        Process a single chat completely. Extract and then fully format it.
-
-    process_list(chat_list: list):
-        Process a list of chats completely.
+        Returns a list of all the sub-lists that couldn't be processed properly.
 
 """
 
-from pydub import AudioSegment
+import concurrent.futures
 from datetime import datetime
-from zipfile import ZipFile
-from shutil import copytree
-from glob import glob
-import threading
-import re
 import os
-
-htmlAudioFormats = {'.mp3': 'mpeg', '.ogg': 'ogg', '.wav': 'wav'}  # Dict of HTML accepted audio formats
-formatDict = {'_': 'em', '*': 'strong', '~': 'del'}  # Dict of format chars with their HTML tags
-
-# Tuple of extensions that can be moved without being converted
-nonConversionExtensions = ('jpg', 'png', 'webp', 'gif', 'mp4', 'mp3', 'ogg', 'wav')
-
-cwd = os.getcwd()
-senderName = htmlFileName = outputDir = ''
-
-# RegEx patterns
-
-fullPrefixPattern = re.compile(r'\[(\d{2}/\d{2}/\d{4}, (\d{1,2}:\d{2}:\d{2} [ap]m|\d{2}:\d{2}:\d{2}))] ([^:]+): ((.|\n)+)')
-# Groups: full prefix is 1, time is 2, name is 3, content is 4
-
-groupMetaPrefixPattern = re.compile(r'\[(\d{2}/\d{2}/\d{4}, (\d{1,2}:\d{2}:\d{2} [ap]m|\d{2}:\d{2}:\d{2}))] (.+)')
-# Groups: full prefix is 1, time is 2, content is 3
-
-attachmentPattern = re.compile(r'<attached: (\d{8}-(\w+)-\d{4}-\d{2}-\d{2}-\d{2}-\d{2}-\d{2})(\.\w+)>$')
-# Groups: filename without extension is 1, file type is 2, extension is 3
-
-# Link pattern taken from urlregex.com
-linkPattern = re.compile(r'http[s]?://(?:[a-zA-Z]|[0-9]|[$-_@.&+]|[!*(),]|(?:%[0-9a-fA-F][0-9a-fA-F]))+')
-
-encryptedMessagesNoticePattern = re.compile(
-    r'\[(\d{2}/\d{2}/\d{4}, (\d{1,2}:\d{2}:\d{2} [ap]m|\d{2}:\d{2}:\d{2}))] ([^:]+): '
-    r'Messages and calls are end-to-end encrypted\. No one outside of this chat, not even WhatsApp, can read or listen to them\.$')
+from pydub import AudioSegment
+import re
+import threading
+import shutil
+import zipfile
 
 
 class Message:
@@ -105,45 +53,83 @@ class Message:
         get_date():
             Return the date of the message.
 
-        create_html() -> str:
+        create_html(sender_name: str) -> str:
             Return HTML representation of the Message object.
 
     """
 
-    def __init__(self, original_string: str, group_chat: bool):
-        """Create a Message object and fully format its content according to the type of message it is.
+    html_audio_formats = {'.mp3': 'mpeg', '.ogg': 'ogg', '.wav': 'wav'}  # Dict of HTML accepted audio formats
+
+    # This is a dictionary of patterns and replacement patterns for converting WhatsApp formatting to HTML tags
+    format_dict = {re.compile(r'\b_([^_]+)_\b'): r'<em>\1</em>',
+                   re.compile(r'\*\b([^*]+)\b\*'): r'<strong>\1</strong>',
+                   re.compile(r'~\b([^~]+)\b~'): r'<del>\1</del>',
+                   re.compile(r'```\b([^`]+)\b```'): r'<code>\1</code>'}
+
+    # Tuple of extensions that can be moved without being converted
+    non_conversion_extensions = ('jpg', 'png', 'webp', 'gif', 'mp4', 'mp3', 'ogg', 'wav')
+
+    # RegEx patterns
+
+    full_prefix_pattern = re.compile(
+        r'\[(\d{2}/\d{2}/\d{4}, (\d{1,2}:\d{2}:\d{2} [ap]m|\d{2}:\d{2}:\d{2}))] ([^:]+): ((.|\n)+)')
+    # Groups: full prefix is 1, time is 2, name is 3, content is 4
+
+    group_meta_prefix_pattern = re.compile(r'\[(\d{2}/\d{2}/\d{4}, (\d{1,2}:\d{2}:\d{2} [ap]m|\d{2}:\d{2}:\d{2}))] ([^:]+)')
+    # Groups: full prefix is 1, time is 2, content is 3
+
+    attachment_message_pattern = re.compile(r'<attached: (\d{8}-(\w+)-\d{4}-\d{2}-\d{2}-\d{2}-\d{2}-\d{2})(\.\w+)>$')
+    # Groups: filename without extension is 1, file type is 2, extension is 3
+
+    # Link pattern taken from urlregex.com
+    link_pattern = re.compile(r'http[s]?://(?:[a-zA-Z]|[0-9]|[$-_@.&+]|[!*(),]|(?:%[0-9a-fA-F][0-9a-fA-F]))+')
+
+    encrypted_messages_notice_pattern = re.compile(
+        r'\[(\d{2}/\d{2}/\d{4}, (\d{1,2}:\d{2}:\d{2} [ap]m|\d{2}:\d{2}:\d{2}))] ([^:]+): '
+        r'Messages and calls are end-to-end encrypted\. No one outside of this chat, not even WhatsApp, can read or listen to them\.$')
+
+    def __init__(self, original_string: str, group_chat: bool, html_file_name: str):
+        """Create a Message object.
 
         Arguments:
-            original_string:
-                The complete original message including all prefix data.
+            original_string: str:
+                The original full content of the message, including all the prefix data in square brackets.
 
-            group_chat:
-                A boolean to determine whether the message is from a group chat.
+            group_chat: bool:
+                A boolean representing whether the message came from a group chat.
+
+            html_file_name: str:
+                The name of the final HTML file.
 
         """
         self._group_chat = group_chat
-        original = original_string
+        self._html_file_name = html_file_name
 
-        try:
-            prefix_match = re.match(fullPrefixPattern, original)
-            self._prefix = prefix_match.group(1)
+        # Remove LRM, LRE, and PDF Unicode characters from original_string
+        original = original_string.replace('\u200e', '').replace('\u202a', '').replace('\u202c', '')
 
+        prefix_match = re.match(Message.full_prefix_pattern, original)
+
+        if prefix_match:  # If it's a normal message
             self._name = prefix_match.group(3)
-            self._content = prefix_match.group(4)
+            self._message_content = prefix_match.group(4)
 
-            if re.match(attachmentPattern, self._content):
-                self._content = add_attachments(self._content)
+            if re.match(Message.attachment_message_pattern, self._message_content):
+                self._format_attachment_message()
             else:
-                self._content = fully_format_message_content(self._content)
+                self._clean_message_content()
+                self._format_with_html_tags()
+                self._format_links()
+
+                self._message_content = self._message_content.replace('\n', '<br>\n\t\t')
 
             self._group_chat_meta = False
-
-        except AttributeError:  # If no match was found for fullPrefixPattern, it's a group chat meta message
-            prefix_match = re.match(groupMetaPrefixPattern, original)
-            self._prefix = prefix_match.group(1)
+        else:  # If it's a group chat meta message
+            prefix_match = re.match(Message.group_meta_prefix_pattern, original)
 
             self._name = ''
-            self._content = clean_html(prefix_match.group(3))
+            self._message_content = prefix_match.group(3)
+            self._clean_message_content()
 
             self._group_chat_meta = True
 
@@ -151,402 +137,374 @@ class Message:
         time = prefix_match.group(2)
 
         # Use a ternary operator to get datetime object whether the message is in 12 hour or 24 hour format
-        date_obj = datetime.strptime(date_raw, '%d/%m/%Y, %I:%M:%S %p') if ' am' in time or ' pm' in time \
+        self._datetime_obj = datetime.strptime(date_raw, '%d/%m/%Y, %I:%M:%S %p') if ' am' in time or ' pm' in time \
             else datetime.strptime(date_raw, '%d/%m/%Y, %H:%M:%S')
 
-        day = datetime.strftime(date_obj, '%d')
-        if day.startswith('0'):
-            day = day.replace('0', '', 1)  # Remove a single leading 0
+        day = str(self._datetime_obj.day)
 
-        self._date = datetime.strftime(date_obj, f'%a {day} %B %Y')
-        self._time = datetime.strftime(date_obj, '%I:%M:%S %p')
+        # Get day of the month extension
+        if day.endswith('1') and day != '11':
+            extension = 'st'
+        elif day.endswith('2') and day != '12':
+            extension = 'nd'
+        elif day.endswith('3') and day != '13':
+            extension = 'rd'
+        else:
+            extension = 'th'
+
+        self._day = day + extension
+
+        self.date = datetime.strftime(self._datetime_obj, f'%a {self._day} %B %Y')
+        self._time = datetime.strftime(self._datetime_obj, '%I:%M:%S %p')
 
         if self._time.startswith('0'):
             self._time = self._time.replace('0', '', 1)
 
     def __repr__(self):
-        """Return a repr of the Message object including the name, date, name, and whether it's from a group chat. Also includes the memory location in hex."""
+        """Return a __repr__ of the Message instance including the name, date, name, and whether it's from a group chat. Also includes the memory location in hex."""
         # Use hex here at end to give memory location of Message object
-        if not self._group_chat_meta:
-            return f'<{self.__class__.__module__}.{self.__class__.__name__} object with name="{self._name}", ' \
-                   f'date="{self._date}", time="{self._time}", and group_chat={self._group_chat} at {hex(id(self))}>'
-        else:
-            return f'<{self.__class__.__module__}.{self.__class__.__name__} object with date="{self._date}", ' \
+        if self._group_chat_meta:
+            return f'<{self.__class__.__module__}.{self.__class__.__name__} object with date="{self.date}", ' \
                    f'time="{self._time}", and group_chat={self._group_chat}, which is a meta message at ' \
                    f'{hex(id(self))}>'
+        else:
+            return f'<{self.__class__.__module__}.{self.__class__.__name__} object with name="{self._name}", ' \
+                   f'date="{self.date}", time="{self._time}", and group_chat={self._group_chat} at {hex(id(self))}>'
 
-    def get_date(self):
-        """Return the date of the message."""
-        return self._date
+    def _clean_message_content(self):
+        """Replace < and > in self._message_content to avoid rogue HTML tags."""
+        self._message_content = self._message_content.replace('<', '&lt;').replace('>', '&gt;')
 
-    def create_html(self) -> str:
-        """Return HTML representation of the Message object."""
+    def _format_with_html_tags(self):
+        """Replace WhatsApp format characters with their HTML tag equivalents in self._message_content."""
+        for pattern, replacement in Message.format_dict.items():
+            self._message_content = re.sub(pattern, replacement, self._message_content)
+
+    def _format_links(self):
+        """Find all the links in self._message_content and wrap them in <a> tags."""
+        links = re.findall(Message.link_pattern, self._message_content)
+
+        for link in links:
+            # Get rid of punctuation or spaces at the end of the link
+            while re.search(r'[\s.,!?]$', link):
+                link = link[:-1]
+
+            formatted_link = f'<a href="{link}" target="_blank">{link}</a>'
+            self._message_content = self._message_content.replace(link, formatted_link)
+
+    def _format_attachment_message(self):
+        """Format an attachment message to properly link to the attachment with HTML tags."""
+        match = re.match(Message.attachment_message_pattern, self._message_content)
+        filename_no_ext = match.group(1)
+        file_type = match.group(2)
+        extension = match.group(3)
+
+        filename = filename_no_ext + extension
+
+        if file_type == 'AUDIO':
+            for ext, given_format in Message.html_audio_formats.items():
+                # If it's a standard, accepted extension, use that format
+                if extension == ext:
+                    html_format = given_format
+                    break
+
+            else:  # If none of the standard HTML audio formats broke out of the for loop, it will be converted to an mp3 file, so use that format
+                html_format = 'mpeg'
+                filename = filename_no_ext + '.mp3'
+
+            self._message_content = f'<audio controls>\n\t\t\t<source ' \
+                                    f'src="Attachments/{self._html_file_name}/{filename}" ' \
+                                    f'type="audio/{html_format}">\n\t\t</audio>'
+
+        elif file_type == 'VIDEO':
+            self._message_content = f'<video controls>\n\t\t\t<source ' \
+                                    f'src="Attachments/{self._html_file_name}/{filename}">\n\t\t</video>'
+
+        elif (file_type == 'PHOTO') or (file_type == 'GIF' and extension == '.gif') or (file_type == 'STICKER'):
+            self._message_content = f'<img class="small" src="Attachments/{self._html_file_name}/{filename}" ' \
+                                    f'alt="IMAGE ATTACHMENT" style="max-height: 400px; max-width: 800px; display: inline-block;">'
+
+        elif file_type == 'GIF' and extension != '.gif':  # Add gif as video that autoplays and loops like a proper gif
+            self._message_content = f'<video autoplay loop muted playsinline>\n\t\t\t<source ' \
+                                    f'src="Attachments/{self._html_file_name}/{filename}">\n\t\t</video>'
+
+        else:
+            self._message_content = f'UNKNOWN ATTACHMENT "{filename}"'
+
+    def create_html(self, sender_name: str) -> str:
+        """Return HTML representation of the Message object.
+
+        Arguments:
+            sender_name: str:
+                The sender in the chat that this message is from.
+
+        """
         if not self._group_chat_meta:
-            if self._name == senderName:
+            if self._name == sender_name:
                 sender_type = 'sender'
             else:
                 sender_type = 'recipient'
 
             # If this is a group chat and this isn't the sender, add the recipient's name
-            if self._group_chat and self._name != senderName:
+            if self._group_chat and self._name != sender_name:
                 css_formatted_name = self._name.replace('\u00a0', '-')  # Replace no-break space with dash
                 recipient_name = f'<span class="recipient-name {css_formatted_name}">{self._name}</span>'
             else:
                 recipient_name = ''
 
             return f'<div class="message {sender_type}">\n\t{recipient_name}\n\t<span class="message-info date">' \
-                   f'{self._date}</span>\n\t\t<p>{self._content}</p>\n\t<span class="message-info time">' \
+                   f'{self.date}</span>\n\t\t<p>{self._message_content}</p>\n\t<span class="message-info time">' \
                    f'{self._time}</span>\n</div>\n\n'
 
         else:  # If a meta message in a group chat
-            return f'<div class="group-chat-meta">\n\t<span class="message-info date">{self._date}</span>\n\t\t' \
-                   f'<p>{self._content}</p>\n\t<span class="message-info time">{self._time}</span>\n</div>\n\n'
+            return f'<div class="group-chat-meta">\n\t<span class="message-info date">{self.date}</span>\n\t\t' \
+                   f'<p>{self._message_content}</p>\n\t<span class="message-info time">{self._time}</span>\n</div>\n\n'
 
 
-def clean_html(string: str) -> str:
-    """Remove < and > to avoid rogue HTML tags."""
-    string = string.replace('<', '&lt;').replace('>', '&gt;')
-    string = string.replace('\"', '&quot;').replace('\'', '&apos;')
-    return string
+class Chat:
+    """The class for each chat to be formatted. Every instance is a separate chat.
 
-
-def simple_format_to_html(string: str) -> str:
-    """Replace WhatsApp format characters with their HTML tags. Doesn't handle code blocks."""
-    first_tag = True
-    list_string = list(string)
-
-    for char, tag in formatDict.items():
-        if char in string and string.count(char) % 2 == 0:
-            for x, letter in enumerate(list_string):
-                if letter == char:
-                    if first_tag:
-                        list_string[x] = f'<{tag}>'
-                        first_tag = False
-                    else:
-                        list_string[x] = f'</{tag}>'
-                        first_tag = True
-
-    return ''.join(list_string)
-
-
-def complete_format_to_html(string: str) -> str:
-    """Replace WhatsApp format characters with their HTML tags. Handles code blocks."""
-    first_tag = True
-    if '```' in string:
-        string = string.replace('```', '<code>')
-        list_string = list(string)
-        for x, letter in enumerate(list_string):
-            if letter == '<':
-                if first_tag:
-                    first_tag = False
-                else:
-                    list_string[x] = '</'
-                    first_tag = True
-
-        string = ''.join(list_string)
-    else:
-        string = simple_format_to_html(string)
-
-    return string
-
-
-def format_links(string: str) -> str:
-    """Find all the links in the string and put them in <a> tags."""
-    # Get links and concat the RegEx groups into a list of links
-    link_match_list = re.findall(linkPattern, string)
-    link_matches = [''.join(match) for match in link_match_list]
-
-    if link_matches:
-        for link in link_matches:
-            if re.match(r'[^A-Za-z]+(\.[^A-Za-z])+', link) and not re.match(r'(\d+\.){3}\d+', link):
-                continue  # If not proper link but also not IP address, skip
-
-            if not link.startswith('http'):
-                working_link = f'http://{link}'  # Create URLs from non URL links
-            else:
-                working_link = link
-
-            formatted_link = f'<a href="{working_link}" target="_blank">{link}</a>'
-            string = string.replace(link, formatted_link)
-
-    return string
-
-
-def fully_format_message_content(string: str) -> str:
-    """Take message content and fully format it properly.
-
-    Clean the HTML of rogue tags, add HTML formatting tags for italics etc, create <a> tags for links, and add <br> tags.
-    """
-    string = clean_html(string)
-    string = complete_format_to_html(string)
-    string = format_links(string)
-    string = string.replace('\n', '<br>\n\t\t')
-    return string
-
-
-def add_attachments(attachment: str) -> str:
-    """Create HTML tags to embed attachments.
-
-    It creates tags for images, videos, GIFs, and audio.
-    This function automatically converts audio files when necessary, typically when they are .opus encoded voice messages.
-    
-    Arguments:
-        attachment:
-            The original, isolated attachment marker. Only the stuff between angled brackets.
-
-    Returns:
-        The HTML tag for the embedded attachment.
+    Methods:
+        format():
+            Fully extract the zip file and format the chat.
 
     """
-    attachment_match = re.match(attachmentPattern, attachment)
-    filename_no_extension = attachment_match.group(1)
-    file_type = attachment_match.group(2)
-    extension = attachment_match.group(3)
 
-    filename = filename_no_extension + extension
+    attachment_file_pattern = re.compile(r'(\d{8}-(\w+)-\d{4}-\d{2}-\d{2}-\d{2}-\d{2}-\d{2})(\.\w+)$')
+    # Groups: filename without extension is 1, file type is 2, extension is 3
 
-    if file_type == 'AUDIO':
-        for ext, html_format in htmlAudioFormats.items():
-            if extension == ext:
-                attachment = f'<audio controls>\n\t\t\t<source src="Attachments/{htmlFileName}/{filename}" ' \
-                                  f'type="audio/{html_format}">\n\t\t</audio>'
-                break
+    def __init__(self, input_file: str, group_chat: bool, sender_name: str, chat_title: str, html_file_name: str, output_dir: str):
+        """Create a Chat object with instance attributes equal to the arguments passed.
 
-        else:  # If none of the standard HTML audio formats broke out of the for loop, convert to .mp3
-            AudioSegment.from_file(f'temp/{filename}').export(
-                f'temp/{filename_no_extension}.mp3', format='mp3')
-            os.remove(f'temp/{filename}')
-            filename = filename_no_extension + '.mp3'
-            attachment = f'<audio controls>\n\t\t\t' \
-                         f'<source src="Attachments/{htmlFileName}/{filename}" type="audio/mpeg">\n\t\t</audio>'
-            os.rename(f'{cwd}/temp/{filename}', f'{outputDir}/Attachments/{htmlFileName}/{filename}')
+        Arguments:
+            input_file:
+                The original zip file of the exported chat. It must have the '.zip' extension.
 
-    elif file_type == 'VIDEO':
-        attachment = f'<video controls>\n\t\t\t<source src="Attachments/{htmlFileName}/{filename}">\n\t\t</video>'
+            group_chat:
+                A boolean to determine whether the chat is a group chat.
 
-    elif (file_type == 'PHOTO') or (file_type == 'GIF' and extension == '.gif') or (file_type == 'STICKER'):
-        attachment = f'<img class="small" src="Attachments/{htmlFileName}/{filename}" alt="IMAGE ATTACHMENT" ' \
-                          'style="max-height: 400px; max-width: 800px; display: inline-block;">'
+            sender_name:
+                The name of the sender in the chat. Typically the user's WhatsApp alias.
 
-    elif file_type == 'GIF' and extension != '.gif':  # Add gif as video that autoplays and loops like a proper gif
-        attachment = f'<video autoplay loop muted playsinline>\n\t\t\t<source src="Attachments/{htmlFileName}/' \
-                          f'{filename}">\n\t\t</video>'
+            chat_title:
+                The title of the chat. This appears in the top bar and in the browser tab. It is not the file name.
 
-    else:
-        attachment = f'UNKNOWN ATTACHMENT "{filename}"'
+            html_file_name:
+                The intended name of the final HTML file.
 
-    return attachment
+            output_dir:
+                The intended directory for the output. The HTML file, Attachments folder, and Library folder will go here.
 
+        """
+        self._input_file = input_file
+        self._group_chat = group_chat
+        self._sender_name = sender_name
+        self._chat_title = chat_title
+        self._html_file_name = html_file_name
+        self._output_dir = output_dir
 
-def write_text(chat_title: str, group_chat: bool):
-    """Format and write all text in _chat.txt to HTML file.
+        # This is a unique temporary directory for this chat, to allow for multithreading multiple chats
+        # os.path.splitext()[0] is used to remove extensions
+        # os.path.split()[1] is used to just get the name of the zip file, not the absolute path
+        self._temp_directory = f'temp_{os.path.splitext(os.path.split(self._input_file)[1])[0]}_{self._chat_title}_' \
+                               f'{os.path.splitext(self._html_file_name)[0]}'
 
-    This function creates a new HTML file with the global htmlFileName.
-    If that filename already exist, it will be given a number in brackets at the end of the name.
-    It only writes the text of _chat.txt to the HTML file.
-    It should be in a thread running parallel with a thread running move_attachment_files().
+        # Make directories if they don't exist
+        if not os.path.isdir(library_path := os.path.join(self._output_dir, 'Library')):
+            shutil.copytree('Library', library_path)
 
-    Arguments:
-        chat_title:
-            A string which becomes the title of the chat. Not the filename, but the title.
+        if not os.path.isdir(attachments_path := os.path.join(self._output_dir, 'Attachments', self._html_file_name)):
+            os.makedirs(attachments_path)
 
-        group_chat:
-            A boolean to determine whether the chat is a group chat.
+    def _extract_zip(self) -> bool:
+        """Extract the zip file into a temporary directory.
 
-    """
-    date_separator = ''
+        Extract the object's input_file into a unique temporary directory. If the extraction fails, the method returns False and skips the file.
 
-    with open('temp/_chat.txt', encoding='utf-8') as f:
-        # Read file and remove LRM, LRE, and PDF Unicode characters
-        chat_txt = f.read().replace('\u200e', '').replace('\u202a', '').replace('\u202c', '')
+        Returns:
+            True if the zip file was successfully extracted, False if not.
 
-    # Add number to end of file if the file already exists
-    if not os.path.isfile(f'{outputDir}/{htmlFileName}.html'):
-        html_file = open(f'{outputDir}/{htmlFileName}.html', 'w+', encoding='utf-8')
-    else:
-        same_name_number = 1
-        while os.path.isfile(f'{outputDir}/{htmlFileName} ({same_name_number}).html'):
-            same_name_number += 1
-        html_file = open(f'{outputDir}/{htmlFileName} ({same_name_number}).html', 'w+', encoding='utf-8')
+        """
+        try:
+            with zipfile.ZipFile(self._input_file) as zip_file:
+                zip_file.extractall(self._temp_directory)
 
-    with open('start_template.txt', encoding='utf-8') as f:
-        start_template = f.readlines()  # f.readlines() preserves \n characters
+            return True
 
-    # Replace chat_title in start_template
-    for i, line in enumerate(start_template):
-        pos = line.find('%chat_title%')
-        if pos != -1:  # If 'chat_title' is found
-            start_template[i] = line.replace('%chat_title%', chat_title)
+        except OSError:  # If the zip file failed to extract
+            print(f'ERROR: Failed to extract {self._input_file}. It likely does not exist. This chat will be skipped.')
+            return False
 
-    for line in start_template:
-        html_file.write(line)
+    def _write_text(self):
+        """Write the contents of temp/_chat.txt to the output directory."""
+        with open(os.path.join(self._temp_directory, '_chat.txt'), 'r', encoding='utf-8') as f:
+            # Read file and remove LRM, LRE, and PDF Unicode characters
+            chat_txt = f.read().replace('\u200e', '').replace('\u202a', '').replace('\u202c', '')
 
-    # Use a re.sub to place LRMs between each message and then create a list by splitting by LRM
-    chat_txt = re.sub(r'\n\[(?=\d{2}/\d{2}/\d{4}, (\d{1,2}:\d{2}:\d{2} [ap]m|\d{2}:\d{2}:\d{2})])', '\u200e[', chat_txt)
-    chat_txt_list = chat_txt.split('\u200e')
+        # Add number to the end of the filename if the file already exists
+        html_filename_with_directory_no_ext = os.path.join(self._output_dir, self._html_file_name)
+        if not os.path.isfile(html_filename_with_directory_no_ext + '.html'):
+            html_file = open(html_filename_with_directory_no_ext + '.html', 'w+', encoding='utf-8')
+        else:
+            same_name_number = 1
 
-    # ===== MAIN WRITE LOOP
+            while os.path.isfile(html_filename_with_directory_no_ext + f' ({same_name_number}).html'):
+                same_name_number += 1
 
-    for string in chat_txt_list:
-        # If it's the notice that messages are encrypted, skip it
-        if re.match(encryptedMessagesNoticePattern, string):
-            continue
+            html_file = open(html_filename_with_directory_no_ext + f' ({same_name_number}).html', 'w+', encoding='utf-8')
 
-        msg = Message(string, group_chat)
+        start_template = open('start_template.txt', 'r', encoding='utf-8').readlines()  # .readlines() preserves \n characters
 
-        if msg.get_date() != date_separator:
-            date_separator = msg.get_date()
-            html_file.write(f'<div class="date-separator">{date_separator}</div>\n\n')
+        # Replace chat title in start template
+        for line in start_template:
+            if '%chat_title%' in line:
+                line = line.replace('%chat_title%', self._chat_title)
 
-        html_file.write(msg.create_html())
+            html_file.write(line)
 
-    with open('end_template.txt', encoding='utf-8') as f:
-        end_template = f.readlines()
+        # Use a re.sub to place LRMs between each message and then create a list by splitting by LRM
+        chat_txt = re.sub(r'\n\[(?=\d{2}/\d{2}/\d{4}, (\d{1,2}:\d{2}:\d{2} [ap]m|\d{2}:\d{2}:\d{2})])', '\u200e[', chat_txt)
+        chat_txt_list = chat_txt.split('\u200e')
 
-    for line in end_template:
-        html_file.write(line)
+        date_separator = ''
 
-    html_file.close()
+        # === Write every message
 
-    os.remove('temp/_chat.txt')
+        for raw_message in chat_txt_list:
+            # If it's the notice that messages are encrypted, skip it
+            if re.match(Message.encrypted_messages_notice_pattern, raw_message):
+                continue
 
+            msg = Message(raw_message, self._group_chat, self._html_file_name)
 
-def move_attachment_files():
-    """Move all attachment files that don't need to be converted.
+            if msg.date != date_separator:
+                date_separator = msg.date
+                html_file.write(f'<div class="date-separator">{date_separator}</div>\n\n')
 
-    This function moves all attachment files that don't need to be converted to the global outputDir.
-    It only moves the attachment files.
-    It should be in a thread running parallel with a thread running write_text().
-    """
-    files = glob('temp/*')
+            html_file.write(msg.create_html(self._sender_name))
 
-    if files:
+        end_template = open('end_template.txt', 'r', encoding='utf-8')
+
+        for line in end_template:
+            html_file.write(line)
+
+        html_file.close()
+
+        os.remove(os.path.join(self._temp_directory, '_chat.txt'))
+
+    def _move_attachment_files(self):
+        """Move the attachment files in temp to the output directory."""
+        files = os.listdir(self._temp_directory)
         for f in files:
-            f = f.replace('\\', '/')
-            filename = f.split('/')[-1]
-            extension = filename.split('.')[-1]
+            if f != '_chat.txt':
+                try:
+                    # Get the file type and the name without an extension
+                    file_type = re.match(Chat.attachment_file_pattern, f).group(2)
+                    f_no_ext = os.path.splitext(f)[0]
 
-            if extension in nonConversionExtensions:
-                os.rename(f'{cwd}/temp/{filename}', f'{outputDir}/Attachments/{htmlFileName}/{filename}')
+                    if file_type == 'AUDIO':
+                        # Convert audio files that can't be played in browsers with simple HTML audio tags
+                        # This is necessary because all voice messages are .opus, which must be converted
+                        if file_type not in Message.html_audio_formats.keys():
+                            # Convert old audio file into .mp3 in same directory
+                            AudioSegment.from_file(os.path.join(self._temp_directory, f)).export(
+                                os.path.join(self._temp_directory, f_no_ext) + '.mp3', format='mp3')
 
+                            # Remove old audio file
+                            os.remove(os.path.join(self._temp_directory, f))
+                            # Set f to new file to make moving the new file easier
+                            f = f_no_ext + '.mp3'
+                except AttributeError:  # File is named weird and doesn't match the RegEx
+                    pass
 
-def extract_zip(input_file: str) -> bool:
-    """Extract the zip file given by file_dir into the temp directory. Return True if successfully extracted.
+                os.rename(os.path.join(self._temp_directory, f), os.path.join(self._output_dir, 'Attachments', self._html_file_name, f))
 
-    If input_file is not a valid zip file, it is skipped and the function returns False.
+    def _start_formatting_threads(self):
+        """Start two threads to fully format the chat after it's been extracted."""
+        self._write_text_thread = threading.Thread(target=self._write_text)
+        self._move_attachment_files_thread = threading.Thread(target=self._move_attachment_files)
 
-    This function must be called before fully_format_chat() to ensure correct operation.
+        self._write_text_thread.start()
+        self._move_attachment_files_thread.start()
 
-    Arguments:
-        input_file:
-            The name of the original input zip file. It must end in '.zip'.
+    def format(self):
+        """Fully extract the zip file and format the chat."""
+        self._extract_zip()
+        self._start_formatting_threads()
 
-    Returns:
-        True if the file is successfully extracted, but False if it isn't.
+        # Wait until directory is empty, then rmdir it
+        while len(os.listdir(self._temp_directory)) > 0:
+            pass
 
-    """
-    try:
-        zip_file_object = ZipFile(input_file)
-        zip_file_object.extractall('temp')
-        zip_file_object.close()
-        return True
-    except OSError:
-        print(f'ERROR: {input_file} was not found. Skipping.')
-        return False
-
-
-def fully_format_chat(group_chat: bool, sender_name: str, chat_title: str, html_file_name: str, output_dir: str):
-    """Fully format an already extracted chat.
-
-    This function will format every message in _chat.txt and write it to output_dir/name.html.
-    It will also move every attachment file to output_dir/Attachments/sender_name.
-
-    extract_zip() must be called before this function to ensure correct operation.
-
-    Arguments:
-        group_chat:
-            A boolean to determine whether the chat is a group chat.
-
-        sender_name:
-            The name of the sender in the chat. Typically the user's WhatsApp alias.
-
-        chat_title:
-            The title of the chat. This appears in the top bar and in the browser tab. It is not the file name.
-
-        html_file_name:
-            The intended name of the final HTML file.
-
-        output_dir:
-            The intended directory for the output. The HTML file, Attachments folder, and Library folder will go here.
-    """
-    global senderName, htmlFileName, outputDir
-
-    senderName = sender_name
-    htmlFileName, _ = os.path.splitext(html_file_name)  # Remove possible '.html' from the end of the file name
-    outputDir = output_dir
-
-    if not os.path.isdir(f'{outputDir}/Library'):
-        copytree('Library', f'{outputDir}/Library')
-
-    if not os.path.isdir(f'{outputDir}/Attachments'):
-        os.mkdir(f'{outputDir}/Attachments')
-
-    if not os.path.isdir(f'{outputDir}/Attachments/{htmlFileName}'):
-        os.mkdir(f'{outputDir}/Attachments/{htmlFileName}')
-
-    text_thread = threading.Thread(target=write_text, args=[chat_title, group_chat])
-    file_move_thread = threading.Thread(target=move_attachment_files)
-    text_thread.start()
-    file_move_thread.start()
-    text_thread.join()
-    file_move_thread.join()
+        os.rmdir(self._temp_directory)
 
 
-def process_single_chat(input_file: str, group_chat: bool, sender_name: str, chat_title: str, html_file_name: str, output_dir: str):
-    """Process a single chat completely. Extract it and then fully format it.
+def process_chat(input_file: str, group_chat: bool, sender_name: str, chat_title: str, html_file_name: str, output_dir: str) -> None:
+    """Process one chat completely.
+
+    This function also checks that all arguments are of the right type before using them. If they're not, raise TypeError.
 
     Arguments:
-        input_file:
+        input_file: str:
             The original zip file of the exported chat. It must have '.zip' on the end.
 
-        group_chat:
+        group_chat: bool:
             A boolean to determine whether the chat is a group chat.
 
-        sender_name:
+        sender_name: str:
             The name of the sender in the chat. Typically the user's WhatsApp alias.
 
-        chat_title:
+        chat_title: str:
             The title of the chat. This appears in the top bar and in the browser tab. It is not the file name.
 
-        html_file_name:
+        html_file_name: str:
             The intended name of the final HTML file.
 
-        output_dir:
+        output_dir: str:
             The intended directory for the output. The HTML file, Attachments folder, and Library folder will go here.
 
     Raises:
-        OSError:
-            If the zip file is not found.
+        TypeError:
+            If the arguments aren't all of the correct type.
 
     """
-    if extract_zip(input_file):
-        fully_format_chat(group_chat, sender_name, chat_title, html_file_name, output_dir)
+    args = [input_file, group_chat, sender_name, chat_title, html_file_name, output_dir]
+
+    arg_types = [type(x) for x in args]
+    printable_arg_types = '[' + ', '.join([str(x)[8:-2] for x in arg_types]) + ']'
+
+    required_types = [str, bool, str, str, str, str]
+    printable_required_types = '[' + ', '.join([str(x)[8:-2] for x in required_types]) + ']'
+
+    # If all the arguments are of the correct type, format the chat
+    if arg_types == required_types:
+        chat = Chat(*args)
+        chat.format()
     else:
-        raise OSError(f'Zip file "{input_file}" not found')
+        raise TypeError(f'Expected arg types of {printable_required_types}. Got {printable_arg_types} instead.')
 
 
-def process_list(chat_list: list):
-    """Process a list of chats completely.
+def process_list_of_chats(list_of_chats: list) -> list:
+    """Fully format a list of lists, where each sub-list is a list of arguments to be passed to process_chat().
 
-    Arguments:
-        chat_list:
-            A list of lists. Each sublist must have 6 elements. These elements are the arguments to be passed to process_single_chat().
-            chat_list should look like this:
-                [[input_file, group_hat, sender_name, chat_title, html_file_name, output_dir],
-                [input_file, group_hat, sender_name, chat_title, html_file_name, output_dir], ...]
+    Returns:
+        rejected_chats:
+            A list of all the sub-lists that couldn't be processed properly. It is an empty list if no sub-lists failed.
 
     """
-    for chat_data in chat_list:
-        # Check if chat_data is the right length and then unpack its values to pass to process_single_chat()
-        if len(chat_data) == 6:
-            process_single_chat(*chat_data)
+    rejected_chats = []
+
+    with concurrent.futures.ThreadPoolExecutor() as executor:
+        # Create a dictionary with the Future object of the method call as the key and the list of args as the value
+        # This allows us to return the args of the rejected chats
+        futures = {executor.submit(process_chat, *chat_data): chat_data for chat_data in list_of_chats}
+
+        for future in concurrent.futures.as_completed(futures):
+            try:
+                future.result()
+            except TypeError:
+                # Get the value from the dictionary using the Future object as the key
+                # This is the arguments passed
+                rejected_chats.append(futures[future])
+
+    return rejected_chats
